Crazy hack to avoid script timeout

+++ web/exportToST.php
@@ -1,8 +1,7 @@
 <?php 
 	include 'dbopen.php';
 	
-	$query = "SELECT at.type as annotationtype, tsi.a2output as a2output, s.filename as filename FROM annotations a, annotationtypes at, tagsetinfos tsi, sentences s WHERE a.annotationtypeid=at.annotationtypeid AND a.tagsetid=tsi.tagsetid AND s.sentenceid=tsi.sentenceid";
-	$result = mysqli_query($con,$query);
+	$step = 100;
 	
 	$basename = 'export';
 	$directory = 'output';
@@ -10,10 +9,42 @@
 	$tarArchive = "$basename.tar";
 	$gzArchive = "$basename.tar.gz";
 	
-	if (file_exists($tarArchive))
-		unlink($tarArchive);
-	if (file_exists($gzArchive))
-		unlink($gzArchive);
+	if (!isset($_GET['start']))
+	{
+	
+		$message = "Starting export...";
+	
+		$query = "SELECT at.type as annotationtype, tsi.a2output as a2output, s.filename as filename, s.sentenceid as sentenceid FROM annotations a, annotationtypes at, tagsetinfos tsi, sentences s WHERE a.annotationtypeid=at.annotationtypeid AND a.tagsetid=tsi.tagsetid AND s.sentenceid=tsi.sentenceid ORDER BY s.sentenceid LIMIT $step";
+		
+		if (file_exists($tarArchive))
+			unlink($tarArchive);
+	}
+	else 
+	{
+		$start = intval($_GET['start']);
+		$message = "Exported $start files...";
+		
+		$query = "SELECT at.type as annotationtype, tsi.a2output as a2output, s.filename as filename, s.sentenceid as sentenceid FROM annotations a, annotationtypes at, tagsetinfos tsi, sentences s WHERE a.annotationtypeid=at.annotationtypeid AND a.tagsetid=tsi.tagsetid AND s.sentenceid=tsi.sentenceid AND s.sentenceid > $start ORDER BY s.sentenceid LIMIT $step";
+		
+	}
+	//print "<p>$query</p>";
+	$result = mysqli_query($con,$query);
+	
+	$phar = new PharData($tarArchive, 0, null, Phar::TAR);
+	
+	$rowCount = mysqli_num_rows($result);
+	if ($rowCount==0)
+	{
+		// We're done
+		
+		if (file_exists($gzArchive))
+			unlink($gzArchive);
+		$phar->compress(Phar::GZ); # Creates archive.tar.gz
+		
+		header("Location: $gzArchive");
+		exit(0);
+	}
+	
 	
 	//$phar->addEmptyDir($directory);
 	
@@ -24,6 +55,7 @@
 		$annotationtype = $row['annotationtype'];
 		$a2output = $row['a2output'];
 		$filename = $row['filename'];
+		$sentenceid = $row['sentenceid'];
 		
 		if (!in_array($filename,$outData))
 			$outData[$filename] = [];
@@ -40,16 +72,56 @@
 			$outData[$filename][] = $line;
 		}
 	}
+	$maxsentenceid = $sentenceid;
 	
-	
-	$phar = new PharData($tarArchive, 0, null, Phar::TAR);
 	foreach ($outData as $filename => $lines)
 	{
-		$fileData = implode("\n",$lines)."\n";
+		$fileData = implode("\n",$lines);
 		$phar->addFromString("$directory/$filename.a2",$fileData);
 	}
 	
-	$phar->compress(Phar::GZ); # Creates archive.tar.gz
+	#header("Location: exportToST.php?start=$maxsentenceid");
 	
-	header("Location: $gzArchive");
+	$redirectURL = "exportToST.php?start=$maxsentenceid";
+	
+	function redirect($url)
+	{
+		$redirectJS = "<script>setTimeout(function() {  window.location.href = '$url'; }, 1000); </script>";
+		echo $redirectJS;
+	}
+	
+	//echo("Location: exportToST.php?start=$maxsentenceid");
 ?>
+<html>
+<head>
+	<title>Annotate - Setup</title>
+    <link href="css/bootstrap.css" rel="stylesheet">
+	<style media="screen" type="text/css">
+
+	.container {
+		position: absolute;
+		top: 50%;
+		left: 50%;
+		transform: translateX(-50%) translateY(-50%);
+		text-align:center;
+	}
+
+	</style>
+
+<?php
+	redirect($redirectURL);
+?>
+	
+</head>
+
+<body>
+
+<div class="container">
+	<p id="text"><?php echo $message ?></p>
+	<p><img id="feedback" src="waiting.gif" /></p>
+	<p><div id="button" style="display:none;"><button class="btn btn-danger" onclick="location.href='admin.php';">Back</button></div></p>
+</div>
+
+
+</body>
+</html>
